feat(lsp): improve the user experience of the analyzer in the editor
